+++ processes/test_lib_functions/test_connection_pg.py
@@ -1,0 +1,3 @@
+from lib.connection_pg import pg_getConnection_via_inifile
+
+conn = pg_getConnection_via_inifile("../..//config/connection_pg_dev.ini","willibald_source")